requirements.txt
Django>=3.2,<5.0
djangorestframework>=3.12
django-cors-headers
django-filter
djangorestframework-simplejwt
drf-yasg
